Separated all DI modules.
